chore: remove LightningBuilder, add LightningV2 Manager trait
