+++ fedimint-testing/src/lib.rs
@@ -10,7 +10,6 @@
 pub mod btc;
 pub mod federation;
 pub mod fixtures;
-pub mod gateway;
 pub mod ln;
 pub use fedimint_testing_core::{db, envs};
 pub use ln_gateway::config::LightningModuleMode;
