+++ gateway/fedimint-lightning/src/lnd.rs
@@ -7,7 +7,6 @@ use std::time::Duration;
 use anyhow::ensure;
 use async_trait::async_trait;
 use bitcoin::hashes::{sha256, Hash};
-use fedimint_core::db::Database;
 use fedimint_core::task::{sleep, TaskGroup};
 use fedimint_core::{secp256k1, Amount, BitcoinAmountOrAll};
 use fedimint_ln_common::contracts::Preimage;
@@ -50,7 +49,7 @@ use crate::{
     CloseChannelsWithPeerRequest, CloseChannelsWithPeerResponse, CreateInvoiceRequest,
     CreateInvoiceResponse, GetBalancesResponse, GetLnOnchainAddressResponse, GetNodeInfoResponse,
     GetRouteHintsResponse, InterceptPaymentRequest, InterceptPaymentResponse, InvoiceDescription,
-    OpenChannelResponse, PayInvoiceResponse, PaymentAction, SendOnchainRequest,
+    LightningV2Manager, OpenChannelResponse, PayInvoiceResponse, PaymentAction, SendOnchainRequest,
     SendOnchainResponse,
 };
 
@@ -65,8 +64,8 @@ pub struct GatewayLndClient {
     tls_cert: String,
     macaroon: String,
     lnd_sender: Option<mpsc::Sender<ForwardHtlcInterceptResponse>>,
-    gateway_db: Database,
     payment_hashes: Arc<RwLock<BTreeSet<Vec<u8>>>>,
+    lnv2_manager: Arc<dyn LightningV2Manager>,
 }
 
 impl GatewayLndClient {
@@ -75,7 +74,7 @@ impl GatewayLndClient {
         tls_cert: String,
         macaroon: String,
         lnd_sender: Option<mpsc::Sender<ForwardHtlcInterceptResponse>>,
-        gateway_db: Database,
+        lnv2_manager: Arc<dyn LightningV2Manager>,
     ) -> Self {
         info!(
             "Gateway configured to connect to LND LnRpcClient at \n address: {},\n tls cert path: {},\n macaroon path: {} ",
@@ -86,7 +85,7 @@ impl GatewayLndClient {
             tls_cert,
             macaroon,
             lnd_sender,
-            gateway_db,
+            lnv2_manager,
             payment_hashes: Arc::new(RwLock::new(BTreeSet::new())),
         }
     }
@@ -292,24 +291,16 @@ impl GatewayLndClient {
                     .await
                     .contains(&payment_hash);
 
-                /*
-                FIXME
                 let db_contains_payment_hash = self_copy
-                    .gateway_db
-                    .begin_transaction_nc()
-                    .await
-                    .load_registered_incoming_contract(PaymentImage::Hash(
-                        sha256::Hash::from_byte_array(
-                            payment_hash
-                                .clone()
-                                .try_into()
-                                .expect("Malformatted payment hash"),
-                        ),
-                    ))
-                    .await
-                    .is_some();
-                    */
-                let db_contains_payment_hash = false;
+                    .lnv2_manager
+                    .contains_incoming_contract(PaymentImage::Hash(sha256::Hash::from_byte_array(
+                        payment_hash
+                            .clone()
+                            .try_into()
+                            .expect("Malformatted payment hash"),
+                    )))
+                    .await;
+
                 let contains_payment_hash = created_payment_hash || db_contains_payment_hash;
 
                 debug!(
@@ -958,7 +949,7 @@ impl ILnRpcClient for GatewayLndClient {
             tls_cert: self.tls_cert.clone(),
             macaroon: self.macaroon.clone(),
             lnd_sender: Some(lnd_sender.clone()),
-            gateway_db: self.gateway_db.clone(),
+            lnv2_manager: self.lnv2_manager.clone(),
             payment_hashes: self.payment_hashes.clone(),
         });
         Ok((Box::pin(ReceiverStream::new(gateway_receiver)), new_client))
